+++ packages/reactter_devtools_extension/lib/src/nodes/dart/plain_instance_node.dart
@@ -20,7 +20,7 @@ base class PlainInstanceNode extends AsyncNode {
     required String key,
     required InstanceRef instanceRef,
   }) {
-    return Rt.createState(
+    return Rt.registerState(
       () => PlainInstanceNode._(
         key: key,
         instanceRef: instanceRef,